Reduce public surface area, unify indentation
There was a side-effect in PR #1 that I noticed after merging: CURLResponse at this point in time should not be public, as none of its functionality is exposed (especially not for successful requests). To undo the leak, I moved the error into a separate, public struct with a public enum and hid CURLResponse again.